+++ imports/ui/components/RecipeCards.jsx
@@ -2,14 +2,14 @@ import React, { Component } from 'react';
 import { withTracker } from 'meteor/react-meteor-data';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
-import Recipes from '../api/recipes';
+import Recipes from '../../api/recipes';
 import { Grid, Paper, Typography } from "@material-ui/core";
 import Card from "@material-ui/core/Card";
 import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
 import CardMedia from "@material-ui/core/CardMedia";
 import Button from "@material-ui/core/Button";
-import { setRecipeDetails } from './actions';
+import { setRecipeDetails } from '../actions';
 import RecipeDetails from './RecipeDetails'
 
 class RecipeCards extends Component {
